Fix signed overflow UB in test_mpsort radix function
radix_int computed the sort key as a signed int64 addition of
INT64_MIN, which overflows for any negative input. clang 22 at -O3
exploits this UB, producing a non-monotonic radix mapping: the sorted
output was correctly ordered by the (miscompiled) radix but not by
integer value, so the test failed with "Ordering of local array is
broken" in the 64-bit case. Do the addition in unsigned arithmetic,
which gives the intended sign-bit flip with defined behavior.

Only the test was affected; the mpsort library code is unchanged.

Co-authored by: Claude Fable 5

+++ libgadget/tests/test_mpsort.cpp
@@ -21,7 +21,7 @@ struct BaseGroup {
 };
 
 static void radix_int(const void * ptr, void * radix, void * arg) {
-    *(uint64_t*)radix = *(const int64_t*) ptr + INT64_MIN;
+    *(uint64_t*)radix = *(const uint64_t*) ptr + (uint64_t) INT64_MIN;
 }
 
 static int64_t
